Added support for snow to spawn no matter the light level. (#422)
* Added support for snow to spawn no matter the light level.

When the "snowSpawnsInAllLightLevels" config option is turned on, snow can accumulate on all blocks regardless of light level. Torch and Soul Torch blocks can now be snow-logged as well.

I've defaulted this to false.
I've been using this on my multiplayer server with my friends for the past few weeks and it's been working great.

I added this to version 1.20.1 since that's the version I play on. I'm not quite sure how merging into all the branches would work though as I'm not super experienced with open source Github projects.

Please let me know if you have any questions or concerns.

* Plants keep their original biome color when they are snow-logged.

This PR removes the old snow-covered textures in favor of staying true to the original biome colors. It keeps the same snow texture from before, but instead overlays it onto the existing texture for biome continuity.

* Revert "Plants keep their original biome color when they are snow-logged."

This reverts commit 6d6998f072101c0fd0820c86ad07dade53be6275.

* Removed torch and soul torch snow logging.